refactor(dialogue_manager): extract test_support.rs with shared mocks
Step 3/9. Move the universally-shared test infrastructure into
dialogue_manager/test_support.rs: ScriptedBackend, EmptyKnowledge,
CountingStore, CountingLearnerStore + the stub_classifier /
stub_extractor / stub_comprehension / default_subsystems /
subsystems_with_extractor / subsystems_with_comprehension builders +
test_learner / chunk / make_test_session_with_turns. Per-test mocks
(RepeatingBackend, FailingLearnerStore, ConceptCapturingStore) stay
where they are next to their tests.

Visibility tightens to pub(super); tests reach via super::test_support::*.
Trims unused imports from mod tests' use block.

mod.rs: 3190 → 2907 lines.
test_support.rs: new, 332 lines.
