+++ src/crates/primer-pedagogy/src/dialogue_manager/mod.rs
@@ -43,6 +43,9 @@ use crate::prompt_pack::{self, PromptPack};
 mod apply;
 use apply::{apply_assessment, apply_comprehension, apply_extraction, merge_concepts_into_turn};
 
+#[cfg(test)]
+mod test_support;
+
 /// Optional persistence stores for a `DialogueManager`.
 ///
 /// Both fields default to `None` — useful for tests that don't care
@@ -1106,291 +1109,21 @@ impl<'a> DialogueManager<'a> {
 
 #[cfg(test)]
 mod tests {
+    use super::test_support::*;
     use super::*;
     use async_trait::async_trait;
     use chrono::Utc;
     use futures::stream;
-    use primer_classifier::StubEngagementClassifier;
     use primer_core::config::PedagogyConfig;
     use primer_core::error::PrimerError;
     use primer_core::inference::{
         GenerationParams, InferenceBackend, Prompt, TokenChunk, TokenStream,
     };
-    use primer_core::knowledge::{KnowledgeBase, Passage, RetrievalParams};
-    use primer_core::learner::{
-        EngagementState, LearnerModel, LearnerProfile, LearningPreferences,
-    };
+    use primer_core::knowledge::KnowledgeBase;
+    use primer_core::learner::{EngagementState, LearnerModel};
     use primer_extractor::ExtractorSettings;
     use std::sync::Mutex;
     use uuid::Uuid;
-
-    fn stub_classifier() -> Arc<dyn EngagementClassifier> {
-        Arc::new(StubEngagementClassifier::new())
-    }
-
-    fn stub_extractor() -> Arc<dyn ConceptExtractor> {
-        Arc::new(primer_extractor::StubConceptExtractor::new())
-    }
-
-    fn stub_comprehension() -> Arc<dyn primer_comprehension::ComprehensionClassifier> {
-        Arc::new(primer_comprehension::StubComprehensionClassifier::new())
-    }
-
-    /// Default-everything subsystems bundle for tests that don't care
-    /// about the specifics of the classifier/extractor/comprehension.
-    fn default_subsystems() -> DialogueManagerSubsystems {
-        DialogueManagerSubsystems {
-            classifier: stub_classifier(),
-            classifier_settings: ClassifierSettings::default(),
-            extractor: stub_extractor(),
-            extractor_settings: ExtractorSettings::default(),
-            comprehension: stub_comprehension(),
-            comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-        }
-    }
-
-    /// Subsystems bundle for tests that need a specific extractor
-    /// (e.g. scripted concepts) but otherwise default classifier/settings.
-    fn subsystems_with_extractor(
-        extractor: Arc<dyn ConceptExtractor>,
-    ) -> DialogueManagerSubsystems {
-        DialogueManagerSubsystems {
-            classifier: stub_classifier(),
-            classifier_settings: ClassifierSettings::default(),
-            extractor,
-            extractor_settings: ExtractorSettings::default(),
-            comprehension: stub_comprehension(),
-            comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-        }
-    }
-
-    /// Subsystems bundle for tests that need a specific comprehension
-    /// classifier but otherwise default classifier/extractor/settings.
-    #[allow(dead_code)]
-    fn subsystems_with_comprehension(
-        comprehension: Arc<dyn primer_comprehension::ComprehensionClassifier>,
-    ) -> DialogueManagerSubsystems {
-        DialogueManagerSubsystems {
-            classifier: stub_classifier(),
-            classifier_settings: ClassifierSettings::default(),
-            extractor: stub_extractor(),
-            extractor_settings: ExtractorSettings::default(),
-            comprehension,
-            comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-        }
-    }
-
-    /// Test inference backend that emits a pre-configured sequence of stream items.
-    struct ScriptedBackend {
-        // Wrap in Mutex<Option> so we can take ownership in `generate_stream`
-        // even though the trait method takes `&self`.
-        script: Mutex<Option<Vec<Result<TokenChunk>>>>,
-        // Counts calls to `summarize` for tests that assert on cadence.
-        summarize_calls: Mutex<u32>,
-    }
-
-    impl ScriptedBackend {
-        fn new(items: Vec<Result<TokenChunk>>) -> Self {
-            Self {
-                script: Mutex::new(Some(items)),
-                summarize_calls: Mutex::new(0),
-            }
-        }
-        fn summary_call_count(&self) -> u32 {
-            *self.summarize_calls.lock().unwrap()
-        }
-        fn set_script(&self, items: Vec<Result<TokenChunk>>) {
-            *self.script.lock().unwrap() = Some(items);
-        }
-    }
-
-    #[async_trait]
-    impl InferenceBackend for ScriptedBackend {
-        fn name(&self) -> &str {
-            "scripted-test"
-        }
-        async fn is_available(&self) -> bool {
-            true
-        }
-        async fn generate_stream(
-            &self,
-            _prompt: &Prompt,
-            _params: &GenerationParams,
-        ) -> Result<TokenStream> {
-            let items = self
-                .script
-                .lock()
-                .unwrap()
-                .take()
-                .expect("ScriptedBackend script already consumed");
-            Ok(Box::pin(stream::iter(items)))
-        }
-        async fn summarize(&self, turns: &[Turn], _target_chars: usize) -> Result<String> {
-            *self.summarize_calls.lock().unwrap() += 1;
-            Ok(format!("[test summary covering {} turns]", turns.len()))
-        }
-    }
-
-    /// Empty knowledge base for tests — never returns any passages.
-    struct EmptyKnowledge;
-    #[async_trait]
-    impl KnowledgeBase for EmptyKnowledge {
-        async fn retrieve(&self, _query: &str, _params: &RetrievalParams) -> Result<Vec<Passage>> {
-            Ok(vec![])
-        }
-    }
-
-    /// Session-store spy: counts `save_session` calls and records the turn
-    /// count of the most recent save. Lets the dialogue-manager tests prove
-    /// the engine actually fired a save (rather than relying on idempotence
-    /// of a manual save after the fact).
-    struct CountingStore {
-        saves: Mutex<u32>,
-        last_turn_count: Mutex<usize>,
-    }
-
-    impl CountingStore {
-        fn new() -> Self {
-            Self {
-                saves: Mutex::new(0),
-                last_turn_count: Mutex::new(0),
-            }
-        }
-        fn save_count(&self) -> u32 {
-            *self.saves.lock().unwrap()
-        }
-        fn last_turn_count(&self) -> usize {
-            *self.last_turn_count.lock().unwrap()
-        }
-    }
-
-    #[async_trait]
-    impl primer_core::storage::SessionStore for CountingStore {
-        async fn save_session(&self, session: &Session) -> Result<()> {
-            *self.saves.lock().unwrap() += 1;
-            *self.last_turn_count.lock().unwrap() = session.turns.len();
-            Ok(())
-        }
-        async fn load_session(&self, _id: uuid::Uuid) -> Result<Option<Session>> {
-            // Stub: tests that need real load behaviour use a different store.
-            Ok(None)
-        }
-        async fn retrieve_session_turns(
-            &self,
-            _session_id: uuid::Uuid,
-            _query: &str,
-            _k: usize,
-            _exclude_indices_at_or_after: usize,
-        ) -> Result<Vec<Turn>> {
-            Ok(vec![])
-        }
-
-        async fn save_classification(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            _turn_index: usize,
-            _assessment: &primer_core::classifier::EngagementAssessment,
-            _classifier_identifier: &str,
-        ) -> Result<()> {
-            Ok(())
-        }
-
-        async fn load_recent_assessments(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            _classifier_identifier: &str,
-            _k: usize,
-        ) -> Result<Vec<primer_core::classifier::EngagementAssessment>> {
-            Ok(vec![])
-        }
-
-        async fn most_recent_session_learner_id(&self) -> Result<Option<uuid::Uuid>> {
-            Ok(None)
-        }
-
-        async fn update_turn_concepts(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            _turn_index: usize,
-            _concepts: &[String],
-        ) -> Result<()> {
-            Ok(())
-        }
-
-        async fn update_exchange_concepts(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            _child_turn_index: usize,
-            _child_concepts: &[String],
-            _primer_turn_index: usize,
-            _primer_concepts: &[String],
-        ) -> Result<()> {
-            Ok(())
-        }
-
-        async fn save_comprehensions(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            _primer_turn_index: usize,
-            _assessments: &[primer_core::comprehension::ComprehensionAssessment],
-            _classifier_identifier: &str,
-        ) -> Result<()> {
-            Ok(())
-        }
-    }
-
-    /// Learner-store spy: counts `save_learner` calls. Used to prove that
-    /// the per-turn save site fires (or doesn't) per the dirty-flag policy.
-    struct CountingLearnerStore {
-        saves: Mutex<u32>,
-    }
-
-    impl CountingLearnerStore {
-        fn new() -> Self {
-            Self {
-                saves: Mutex::new(0),
-            }
-        }
-        fn save_count(&self) -> u32 {
-            *self.saves.lock().unwrap()
-        }
-    }
-
-    #[async_trait]
-    impl primer_core::storage::LearnerStore for CountingLearnerStore {
-        async fn save_learner(&self, _learner: &LearnerModel) -> Result<()> {
-            *self.saves.lock().unwrap() += 1;
-            Ok(())
-        }
-        async fn load_learner(&self) -> Result<Option<LearnerModel>> {
-            Ok(None)
-        }
-    }
-
-    fn test_learner() -> LearnerModel {
-        LearnerModel {
-            profile: LearnerProfile {
-                id: Uuid::new_v4(),
-                name: "Tester".to_string(),
-                age: 8,
-                languages: vec!["en".to_string()],
-                locale: primer_core::i18n::Locale::English,
-                created_at: Utc::now(),
-                last_active: Utc::now(),
-            },
-            concepts: vec![],
-            preferences: LearningPreferences::default(),
-            current_engagement: EngagementState::Engaged,
-            recent_assessments: vec![],
-        }
-    }
-
-    fn chunk(text: &str, done: bool) -> TokenChunk {
-        TokenChunk {
-            text: text.to_string(),
-            done,
-        }
-    }
 
     #[tokio::test]
     async fn respond_to_streaming_invokes_callback_per_chunk() {
@@ -1681,25 +1414,6 @@ mod tests {
     }
 
     // ─── resume_session and summary refresh ──────────────────────────
-
-    fn make_test_session_with_turns(n: usize, learner_id: Uuid) -> Session {
-        use primer_core::conversation::Speaker;
-        let mut session = Session::new(learner_id);
-        for i in 0..n {
-            session.add_turn(Turn {
-                speaker: if i % 2 == 0 {
-                    Speaker::Child
-                } else {
-                    Speaker::Primer
-                },
-                text: format!("turn {i}"),
-                timestamp: Utc::now(),
-                intent: None,
-                concepts: vec![],
-            });
-        }
-        session
-    }
 
     #[tokio::test]
     async fn resume_session_loads_turns_without_greeting() {
